app.py
# Si app.py n'existe pas, copy-paste tout ça:

@"
import streamlit as st
from supabase import create_client, Client
import os
from dotenv import load_dotenv
from datetime import datetime
import time

# Load environment
load_dotenv()

# Page config
st.set_page_config(
    page_title="🥓 BaconAlgo - Premium Trading Platform",
    page_icon="🥓",
    layout="wide"
)

# Bacon theme
st.markdown('''
<style>
    @import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;600;700;900&display=swap');
    
    * { font-family: 'Inter', sans-serif; }
    
    .stApp {
        background: linear-gradient(135deg, #0a0a0a 0%, #1a0f0a 100%);
    }
    
    .main-header {
        background: linear-gradient(135deg, #D2691E, #FF8C00);
        padding: 40px;
        border-radius: 20px;
        text-align: center;
        margin-bottom: 30px;
        box-shadow: 0 10px 40px rgba(210, 105, 30, 0.4);
    }
    
    .main-header h1 {
        color: white;
        font-size: 64px;
        font-weight: 900;
        margin: 0;
        text-shadow: 3px 3px 6px rgba(0,0,0,0.3);
    }
    
    .stButton>button {
        background: linear-gradient(135deg, #D2691E, #FF8C00);
        color: white;
        border: none;
        border-radius: 12px;
        padding: 16px 32px;
        font-weight: 700;
        font-size: 18px;
        transition: all 0.3s ease;
    }
    
    .stButton>button:hover {
        transform: scale(1.05);
        box-shadow: 0 8px 20px rgba(255, 140, 0, 0.5);
    }
</style>
''', unsafe_allow_html=True)

# Main app
st.markdown('''
<div class="main-header">
    <h1>🥓 BACONALGO</h1>
    <p style="color: white; font-size: 24px; margin: 10px 0 0 0;">
        Premium Trading Platform - Live Now! 🔥
    </p>
</div>
''', unsafe_allow_html=True)

col1, col2, col3 = st.columns(3)

with col1:
    st.metric("🔥 Signals Today", "8", "+3")
    
with col2:
    st.metric("💰 Win Rate", "73%", "+5%")
    
with col3:
    st.metric("📊 Active Traders", "1,247", "+89")

st.divider()

st.markdown("## 🚀 Welcome to BaconAlgo!")
st.success("✅ Platform is LIVE and running on Cloudflare + Vercel!")

if st.button("🔍 Start Trading", type="primary", use_container_width=True):
    st.balloons()
    st.success("🎉 Trading dashboard coming soon!")

st.info("🛡️ Protected by Cloudflare | ⚡ Powered by Vercel | 🥓 Made with Bacon")
"@ | Out-File -FilePath app.py -Encoding UTF8

Write-Host "✅ app.py created!" -ForegroundColor Green
