requirements.txt
streamlit==1.29.0
supabase==2.0.3
python-dotenv==1.0.0
